Revert adding netlib_uri
It throws if it fails to parse (a perfectly valid URI), and it can't
distinguish between query strings occurring within a hash fragment or vice
versa. And there's that whole problem with forcing a static C runtime on
MSVC, which is just ridiculous and would bloat the executable.

I guess I didn't really need this thing :)