Do lots of things!
1) move headers from src/ to include/twitchsw/ to get them outta the way

2) implement a WebView to enable the implicit authentication flow for
Twitch.

The WebView needs a lot of work: no cookies or session saving of any kind
just yet, and the API is probably pretty awful. It's what we got for now.

TODO: implement this using webkitgtk+ and UIWebKit on mac/linux without
having to add WebKit as a submodule :)

A good chunk of this C++ webview was stolen from
https://github.com/Tobbe/CbbIEEmbed, so @Tobbe gets a shoutout for helping
make this thing work. ActiveX controls are terrible, and I have half the
mind to just find a way to incorporate the .Net browser component rather
than this ACtiveX shit.

3) The implementation basically works for now on windows, which is really
the only platform OBS Studio does anything useful with at the moment due
to shitty streaming support on mac/linux. So most of the remaining work on
this is cleanup, and usability improvements.