Update (if needed) when streaming begins
Also adds optional mechanism for starting the twitch update process
without streaming (useful for debugging).

To do this, debug with the environment variable
TSW_UPDATE_WITHOUT_STREAMING set to a value that is not "0", "false",
"no", or "off". This may not necessarily work on non-windows platforms,
and may stop working if obs modules are ever initialized outside of the
main thread. But for now, this is golden.