Use WebKit WTF-style smart ptrs rather than the weird solution.
This transforms String into a WTFString-influenced class, with a similar mechanic
for static string management. Similarly, all of the old home-rolled smart pointers
with weird semantics have been replaced with WebKit style smart ptrs with explicit
semantics.

Ref -> assumed to be non-null (dereferencing a nulled out Ref<T> is a
       programmer error, which can only occur after ownership of the reference
       has been given away)

RefPtr -> Like a Ref, but with pointer-like semantics (is nullable)

WeakPtr -> essentially a RefPtr that does not retain what it refers to,
       currently used in the WebView implementation to prevent cycles between
       WebViewImpl and WebView. The WTF WeakPtr implementation is very alloc-happy
       for null WeakPtrs, which is a bit silly. I'd like to find a fix for that.

Not using bmalloc, but it might be worth investigating if Yusuke Suzuki's port
is reasonably up to date (doesn't look like it's been touched in a few years).

Closes #13