Add the "real" mac WebView implementation.
It basically works, although it's not pretty and could use an awful lot of cleanup.

Similar issues exist with it as the windows one, in that it fails to "remember"
login credentials for Twitch. I don't want to save these and use the WebViewController
hook which could circumvent that, so need to find another workaround.

Closes #13